GlobalContext file addition
